Added equations, test and a curve drawer
Added the necessary equations for the interpolations
Tested following an object
Added variants of the 2nd order interpolation equations to that try to keep the system stable (Not grow to infinity)
Added a graph drawing script. It redraws the animation.
Added a manual stop counter to avoid infinite loop in the main graph drawing function, though an infinite loop can still happen in the interpolation function that it calls (may need another variable their).
The script was made to run in the editor, but that idea was cancelled (commented) just to avoid problems in the edit mode itself (Not sure if it really benefits, since an inf loop is still bad in the playmode).
The drawing scripts depends on the original follow script to draw the step response curve based on its values. It can also be easily modified (simple copy paste) to apply the other variants of the interpolation.